requirements.txt
appdirs==1.4.4
biopython==1.79
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
networkx==2.8.7
numpy==1.23.3
pandas==1.5.0
prettytable==3.4.1
pydna==4.0.7
pygenome==3.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
tqdm==4.64.1
urllib3==1.26.12
wcwidth==0.2.5
flake8==5.0.4